feat: fetch web identity credentials
chore: role session name

fix: remove blank line
