+++ common/etc/nginx/include/aws_common.js
@@ -1,4 +1,3 @@
-
 /*
  *  Copyright 2023 F5, Inc.
  *
@@ -403,12 +402,72 @@ async function fetchEC2RoleCredentials() {
     };
 }
 
+/**
+ * Get the credentials by assuming calling AssumeRoleWithWebIdentity with the environment variable
+ * values ROLE_ARN, AWS_WEB_IDENTITY_TOKEN_FILE and HOSTNAME
+ *
+ * @returns {Promise<{accessKeyId: (string), secretAccessKey: (string), sessionToken: (string), expiration: (string)}>}
+ * @private
+ */
+async function fetchWebIdentityCredentials(roleSessionName) {
+    const arn = process.env['AWS_ROLE_ARN'];
+    const name = process.env['HOSTNAME'] || roleSessionName;
+
+    let sts_endpoint = process.env['STS_ENDPOINT'];
+    if (!sts_endpoint) {
+        /* On EKS, the ServiceAccount can be annotated with
+           'eks.amazonaws.com/sts-regional-endpoints' to control
+           the usage of regional endpoints. We are using the same standard
+           environment variable here as the AWS SDK. This is with the exception
+           of replacing the value `legacy` with `global` to match what EKS sets
+           the variable to.
+           See: https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html
+           See: https://docs.aws.amazon.com/eks/latest/userguide/configure-sts-endpoint.html */
+        const sts_regional = process.env['AWS_STS_REGIONAL_ENDPOINTS'] || 'global';
+        if (sts_regional === 'regional') {
+            /* STS regional endpoints can be derived from the region's name.
+               See: https://docs.aws.amazon.com/general/latest/gr/sts.html */
+            const region = process.env['AWS_REGION'];
+            if (region) {
+                sts_endpoint = `https://sts.${region}.amazonaws.com`;
+            } else {
+                throw 'Missing required AWS_REGION env variable';
+            }
+        } else {
+            // This is the default global endpoint
+            sts_endpoint = 'https://sts.amazonaws.com';
+        }
+    }
+
+    const token = fs.readFileSync(process.env['AWS_WEB_IDENTITY_TOKEN_FILE']);
+
+    const params = `Version=2011-06-15&Action=AssumeRoleWithWebIdentity&RoleArn=${arn}&RoleSessionName=${name}&WebIdentityToken=${token}`;
+
+    const response = await ngx.fetch(sts_endpoint + "?" + params, {
+        headers: {
+            "Accept": "application/json"
+        },
+        method: 'GET',
+    });
+
+    const resp = await response.json();
+    const creds = resp.AssumeRoleWithWebIdentityResponse.AssumeRoleWithWebIdentityResult.Credentials;
+
+    return {
+        accessKeyId: creds.AccessKeyId,
+        secretAccessKey: creds.SecretAccessKey,
+        sessionToken: creds.SessionToken,
+        expiration: creds.Expiration,
+    };
+}
+
 export default {
     awsHeaderDate,
     buildCanonicalRequest,
     buildSigningKeyHash,
     eightDigitDate,
     fetchEC2RoleCredentials,
+    fetchWebIdentityCredentials,
     readCredentials,
     securityToken,
     signedHeaders,
